requirements.txt
PyYAML==6.0.3
pandas==2.2.3
numpy==2.2.2
xgboost==2.1.3
scikit-learn==1.5.2
openpyxl==3.1.5
boruta==0.4.3
tqdm==4.67.1